Replace `u64` with `i64` in `wp_mobile_cache` and `wp_mobile` crates (#1032)
Eliminates unnecessary casting between rusqlite's `i64` and our `u64` types.
Since rusqlite consistently uses `i64` for all integer operations, aligning
our types reduces friction and improves code clarity.

Changes:
- Change `RowId` inner type from `u64` to `i64`
- Remove casting in `RowId`'s `ToSql`, `FromSql`, and conversion implementations
- Update migration functions to use `i64` instead of `u64`
- Change `FetchResult.total_items` from `Option<u64>` to `Option<i64>`
- Remove unnecessary casts in repository and service code
- Update type casts in test code and helper functions

+++ wp_mobile_cache/src/repository/term_relationships.rs
@@ -271,7 +271,7 @@ impl TermRelationshipRepository {
             HashMap::new(),
             |mut acc: HashMap<i64, Vec<DbTermRelationship>>, row_result| {
                 let relationship = row_result.map_err(SqliteDbError::from)?;
-                acc.entry(relationship.object_id.0 as i64)
+                acc.entry(relationship.object_id.0)
                     .or_default()
                     .push(relationship);
                 Ok::<_, SqliteDbError>(acc)
